requirements.txt
asgiref==3.5.2
boto3==1.26.11
botocore==1.29.11
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
gunicorn==20.1.0
jmespath==1.0.1
jsonpickle==2.0.0
numpy==1.20.3
opencv-python==4.5.5.64
pandas==1.2.4
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
virtualenv==20.16.5
whitenoise==6.2.0
zss==1.2.0
